test: extract booking test harness
